Make script code more DRY
