+++ scripts/inject-env.ts
@@ -35,4 +35,3 @@ async function main() {
 }
 
 runAsyncProcess(main);
-
